# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemEvidencia": {
        "additionalProperties": false,
        "description": "Item de ``GET /v1/evidencia/{explicacion_id}``.",
        "properties": {
          "fact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fact Id"
          },
          "ref_id": {
            "title": "Ref Id",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "tipo": {
            "description": "\"linea\" | \"mov\" | \"cat\" | \"tramo\" | \"faq\" | \"casuistica\"",
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "tipo",
          "ref_id",
          "snippet"
        ],
        "title": "ItemEvidencia",
        "type": "object"
      },
...
      "RespuestaEvidencia": {
        "additionalProperties": false,
        "description": "Cuerpo de ``GET /v1/evidencia/{explicacion_id}``.",
        "properties": {
          "corpus_saneado": {
            "default": true,
            "description": "Los items de corpus vienen sin cifras, por construcción",
            "title": "Corpus Saneado",
            "type": "boolean"
          },
          "explicacion_id": {
            "title": "Explicacion Id",
            "type": "string"
          },
          "factset_sha256": {
            "title": "Factset Sha256",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemEvidencia"
            },
            "title": "Items",
            "type": "array"
          },
          "periodo": {
            "title": "Periodo",
            "type": "string"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "explicacion_id",
          "trace_id",
          "periodo",
          "factset_sha256"
        ],
        "title": "RespuestaEvidencia",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT HS256 emitido por /dev/token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "concursos-cis@ulima.edu.pe",
      "name": "Equipo recibo-claro"
    },
...
    "license": {
      "name": "MIT"
    },
    "summary": "Explica la variación del recibo con cero invenciones financieras, verificadas en código y comprobables por logs.",
    "title": "recibo-claro · Explicación inteligente de recibos Movistar",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/evidencia/{explicacion_id}": {
      "get": {
        "description": "Devuelve la evidencia del turno identificado por ``explicacion_id``.\n\nEl identificador es el ``trace_id`` del turno, que la respuesta de ``/v1/explicar``\npublica en ``telemetria.explicacion_id`` y en la cabecera ``X-Trace-Id``.",
        "operationId": "obtener_evidencia_v1_evidencia__explicacion_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "explicacion_id",
            "required": true,
            "schema": {
              "title": "Explicacion Id",
              "type": "string"
            }
          },
          {
            "description": "Filtra por tipo: linea, mov, tramo, cat, faq, casuistica",
            "in": "query",
            "name": "solo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por tipo: linea, mov, tramo, cat, faq, casuistica",
              "title": "Solo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaEvidencia"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "EXPLICACION_NO_ENCONTRADA"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evidencia que respalda una explicación entregada",
        "tags": [
          "evidencia"
        ]
      }
    },
...
  },
...
}